DEV: Check if materialized view exists before querying against it (#125)
Why this change?

Previously, we were querying against the materialized view even when it
does not exists and rescues the `ActiveRecord::StatementInvalid` error.
However, `ActiveRecord` relies on that error being raised to decide if a
transaction should be rolled backed. Since we swallow the error, the
transaction ends up not being rolled back and left in an aborted state
which is not ideal.

What does this change do?

We check for the existence of the materialized view before querying
against it to prevent unnecessary errors.