DEV: Remove explicit leaderboard materialize view versions (#187)
Leaderboard materialized views currently include a version number
in their view names. This versioning mechanism helps with staleness
checks and ensures views are updated after changes to the backing query.
While effective, it also means the view name changes with each version bump,
making it unsuitable for use in badge and Data Explorer queries.

Ideally, given the transient nature of these materialized views, they shouldn’t
be used in user queries, but we don’t have much control over that.

This change removes explicit version numbers and instead adds a `COMMENT` to each
view containing a signature of the query at the time it was created.
This approach maintains the ability to check for staleness without explicitly versioning
materialized views.