DEV: Fix failing tests depending on the time of the month
In SQL land, we filter for the scores using `CURRENT_DATE - INTERVAL '1
month'` but in the test setup we are using `Date.current - 30.days`.
These two calculators are different where the former is looking at the date 1
calendar month ago while the latter is just returning the date 30 days
ago.