FIX: Disable solutions scorable when the solved plugin is disabled (#199)
Before https://github.com/discourse/discourse-solved/pull/342, scorables for solved were in custom fields, this means no errors or validations were required since every instance is able to query custom fields.

Now we will have to check for the availability of discourse_solved prior to scoring. This commit disables the scorable if the plugin is not enabled or installed.